Add a LabelValuePanel Component
